Removed the unnecessary didtimeout selector, I will handle timeouts by setting max timeouts to 1 and handling timeouts manually.

+++ Classes/Tests/ASIHTTPRequestTests.m
@@ -554,7 +554,7 @@
 
 	// Cancel the request as soon as it has downloaded 64KB
 	while (1) {
-		sleep((Float32)0.5);
+		sleep(0.5);
 		if ([request totalBytesRead] > 32*1024) {
 			[request cancel];
 			break;
@@ -579,7 +579,7 @@
 	[request startAsynchronous];
 
 	while (1) {
-		sleep((Float32)0.5);
+		sleep(0.5);
 		if ([request isFinished]) {
 			break;
 		}
